@robodoo 2023-10-26
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/139844).